pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aaco"
version = "1.0.0"
description = "AMD AI Compute Observatory - Full-stack observability and performance intelligence for AMD AI workloads"
readme = "README.md"
license = {text = "Proprietary Source-Available"}
requires-python = ">=3.10"
authors = [
    {name = "Sudheer Ibrahim Daniel Devu", email = "sudheeridevu@gmail.com"}
]
maintainers = [
    {name = "Sudheer Ibrahim Daniel Devu", email = "sudheeridevu@gmail.com"}
]
keywords = [
    "amd", "rocm", "gpu", "profiling", "observability",
    "ai", "inference", "performance", "benchmark",
    "onnx", "migraphx", "kernel", "tracing", "performance-analysis",
    "deep-learning", "machine-learning", "pytorch", "ebpf", "healthcare-ai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Hardware",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
onnx = [
    "onnx>=1.14.0",
    "onnxruntime>=1.16.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
]
ml = [
    "scikit-learn>=1.3.0",
    "shap>=0.43.0",
]
ebpf = [
    "bcc>=0.28.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]
all = [
    "aaco[onnx,dashboard,ml,dev,docs]",
]

[project.scripts]
aaco = "aaco.cli:main"

[project.urls]
Homepage = "https://github.com/SID-Devu/AMD-AI-Compute-Observatory"
Repository = "https://github.com/SID-Devu/AMD-AI-Compute-Observatory"
Documentation = "https://sid-devu.github.io/AMD-AI-Compute-Observatory/"
Issues = "https://github.com/SID-Devu/AMD-AI-Compute-Observatory/issues"
Changelog = "https://github.com/SID-Devu/AMD-AI-Compute-Observatory/releases"

[tool.setuptools.packages.find]
where = ["."]
include = ["aaco*"]

[tool.setuptools.package-data]
aaco = ["report/templates/*.j2", "report/templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow running",
    "rocm: marks tests requiring ROCm",
    "integration: marks integration tests",
    "gpu: marks tests requiring GPU",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | sessions
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "build", "dist", "sessions"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["sessions", "build", "dist"]

[tool.coverage.run]
source = ["aaco"]
omit = ["*/tests/*", "*/__pycache__/*", "*/conftest.py"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
fail_under = 70
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = ["tests", "sessions"]
skips = ["B101"]
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [".git", ".venv", "build", "dist", "sessions", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # line-too-long - handled by formatter
    "E722",  # bare-except - common in system/hardware access code
    "E402",  # module-level-import-not-at-top - used for conditional imports
    "E741",  # ambiguous-variable-name - sometimes legitimate
    "F401",  # unused-import - many are for try/except availability checks
    "F403",  # star-imports - used in some modules intentionally
    "F841",  # local-variable-is-assigned-to-but-never-used - informational
    "W291",  # trailing-whitespace - handled by formatter
    "W293",  # blank-line-contains-whitespace - handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"